release 0.02

+++ src/posix/close.c
@@ -1,6 +1,6 @@
 #include <pl-unistd.h>
 
-long close(int fd){
+int close(int fd){
 	void* ptr[1] = { (void*)fd };
 	return pl_syscall(__NR_close, 1, ptr);
